+++ R/globals.R
@@ -3,7 +3,6 @@
 utils::globalVariables(c(
   ".data", # <epidist_diagnostics>
   "samples", # <epidist_diagnostics>
-  ".data", # <.add_default_max_delay>
   ".data", # <as_epidist_latent_model.epidist_linelist_data>
   "woverlap", # <epidist_stancode.epidist_latent_model>
   ".data", # <as_epidist_linelist_data.epidist_aggregate_data>